Add Twitter Login to Navbar
Adds Twitter login functionality. A guest will see a 'Login with Twitter'
button in the navbar and is able to be authenticated with their twitter
account, and the authenticated user is directed to the / page where they
now see a logout button in the navbar. We are still developing the logout
functionality and will then move on to the user dashboard/show page.