Add UsersLegislators model
UsersLegislators model acts as joining model between Users
and Legislators. To get the instances of UsersLegislators
for a user you must have the user object and call
'object.following.all()'. This is a great breakthrough; however,
I think there is likely a better way to access the legislators
a user is following more directly, as the above query will
return an array-like object of UsersLegislators instances.