Add functionality for follow/unfollow legislator
Builds out JS functinos on legislator detail page to follow and unfollow a legislator.
Functions utilize new routes via post requests through which the logged in user id
and the legislator id are passed. The endpoints return a 200 status code and either
establish or break the relationship as necessary. The commit also adds a list-group
to the user_show page which displays each of the legislators name, chamber, state, and
party. Clicking on a legislator will take you to their show page. Follow/un-follow
is functioning as intended at this time.