Modify detail.html
Adds conditional statement that modifies display of follow button.
In the event that the user is already following the legislator,
the unfollow button will appear. In the event that the user
is not following the legislator, the follow button will appear.
All page behavior is appropriate and as desired at this time.