Make navbar accessible throughout application
Extracts navbar css modification function to navbar.js. This file is
added into head of all views. Page behavior is unaffected except for desired
changes in navbar which have been applied throughout the site. I was unable to
find a way to override the bootstrap default settings in the css file, so I
opted for modification with JS.